+++ docs/api/perry.d.ts
@@ -1,6 +1,6 @@
 // Auto-generated from Perry's API manifest (#465). Do not edit by hand.
 // Source: perry-api-manifest::API_MANIFEST
-// Coverage: 1317 entries across 81 modules
+// Coverage: 1319 entries across 81 modules
 
 declare module "@perryts/pdf" {
   /** stdlib */
@@ -2031,9 +2031,13 @@ declare module "validator" {
 
 declare module "worker_threads" {
   /** stdlib */
+  export function getEnvironmentData(p0: any): any;
+  /** stdlib */
   export function getWorkerData(...args: any[]): any;
   /** stdlib */
   export function parentPort(...args: any[]): any;
+  /** stdlib */
+  export function setEnvironmentData(p0: any, p1: any): void;
   /** stdlib */
   export function workerData(...args: any[]): any;
 }